requirements.txt
# Core dependencies for LSTM Sentiment Classifier

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0

# Data Processing
pandas>=1.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
PyYAML>=5.4.0

# Progress Bars
tqdm>=4.62.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utilities
requests>=2.25.0
urllib3>=1.26.0

# Web UI
streamlit>=1.28.0
plotly>=5.15.0

# Optional NLP Libraries for Advanced Preprocessing
# Install with: pip install nltk spacy
# For spaCy: python -m spacy download en_core_web_sm
nltk>=3.6.0
spacy>=3.4.0